Rename QueryBar to QueryBarTopRow (#45250)
* rename QueryBar to QueryBarTopRow

* Fixed test typo

* Fixed search bar mock
