+++ src/legacy/core_plugins/data/public/query/query_bar/components/query_bar_top_row.test.mocks.tsx
