+++ src/legacy/core_plugins/data/public/query/query_bar/index.ts
@@ -17,7 +17,9 @@
  * under the License.
  */
 
-export { QueryBar, QueryBarInput } from './components';
+export { QueryBarTopRow } from './components/query_bar_top_row';
+export { QueryBarInput } from './components/query_bar_input';
+
 export { fromUser } from './lib/from_user';
 export { toUser } from './lib/to_user';
 export { getQueryLog } from './lib/get_query_log';
