+++ src/legacy/core_plugins/data/public/search/search_bar/components/search_bar.tsx
@@ -27,7 +27,8 @@ import { get, isEqual } from 'lodash';
 
 import { toastNotifications } from 'ui/notify';
 import { UiSettingsClientContract, SavedObjectsClientContract } from 'src/core/public';
-import { IndexPattern, Query, QueryBar, FilterBar } from '../../../../../data/public';
+import { IndexPattern, Query, FilterBar } from '../../../../../data/public';
+import { QueryBarTopRow } from '../../../query';
 import { SavedQuery, SavedQueryAttributes } from '../index';
 import { SavedQueryMeta, SaveQueryForm } from './saved_query_management/save_query_form';
 import { SavedQueryManagementComponent } from './saved_query_management/saved_query_management_component';
@@ -366,7 +367,7 @@ class SearchBarUI extends Component<SearchBarProps, State> {
     let queryBar;
     if (this.shouldRenderQueryBar()) {
       queryBar = (
-        <QueryBar
+        <QueryBarTopRow
           uiSettings={this.props.uiSettings}
           savedObjectsClient={this.props.savedObjectsClient}
           query={this.state.query}
